"type : [#feature] 사이드바,푸터 공용컴포넌트 추가"
